Nest function rules under a top-level regex
This means that we won't match something like
```
(arg: int) -> int
```
on its own.

The `end` rule will either run to the end of the document, if this
function is at the bottom of the file, or till the next top-level
language construct. We don't have semicolons at the top-level scope, so
there's not much else we can do.